# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully."
          },
          "409": {
            "description": "Email already exists."
          }
        },
        "summary": "Register a customer account",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Login successful and access token returned."
          },
          "401": {
            "description": "Invalid email or password."
          }
        },
        "summary": "Login with email and password",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Refresh token rotated successfully."
          },
          "401": {
            "description": "Invalid refresh token."
          }
        },
        "summary": "Rotate refresh token and issue a new access token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/google": {
      "post": {
        "operationId": "AuthController_googleLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleLoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Login via Google successful, tokens and user info returned."
          },
          "401": {
            "description": "Invalid Google credential."
          }
        },
        "summary": "Login or register via Google Identity Services (GIS)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Logged out successfully."
          },
          "401": {
            "description": "Invalid active session."
          }
        },
        "summary": "Logout current device session",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/admin/partners/{partnerProfileId}/kyc": {
      "patch": {
        "operationId": "AdminController_updateKycStatus",
        "parameters": [
          {
            "name": "partnerProfileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateKycStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Partner KYC status updated successfully."
          },
          "403": {
            "description": "Admin role is required."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Approve or reject a partner KYC application",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/admin/properties/{propertyId}/status": {
      "patch": {
        "operationId": "AdminController_updatePropertyStatus",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePropertyStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Property moderation status updated successfully."
          },
          "403": {
            "description": "Admin role is required."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Approve, reject, or suspend a property listing",
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/api/bookings/me": {
      "get": {
        "operationId": "BookingsController_findMine",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns bookings for the authenticated customer."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List bookings for the current customer",
        "tags": [
          "Bookings"
        ]
      }
    },
    "/api/bookings/my-bookings": {
      "get": {
        "operationId": "BookingsController_findMyBookings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns bookings for the authenticated customer."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Alias for /bookings/me",
        "tags": [
          "Bookings"
        ]
      }
    },
    "/api/bookings/{id}": {
      "get": {
        "operationId": "BookingsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns booking details."
          },
          "404": {
            "description": "Booking not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get booking details by ID",
        "tags": [
          "Bookings"
        ]
      }
    },
    "/api/bookings/{id}/cancel": {
      "post": {
        "operationId": "BookingsController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Booking cancelled and availability restored."
          },
          "404": {
            "description": "Booking not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel a booking and restore availability",
        "tags": [
          "Bookings"
        ]
      }
    },
    "/api/bookings/{id}/reviews": {
      "post": {
        "operationId": "BookingsController_createReview",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingCreateReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Booking review created."
          },
          "400": {
            "description": "Booking is not eligible for a review."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Review a checked-out booking",
        "tags": [
          "Bookings"
        ]
      }
    },
...
    "/api/partner/bookings/{id}/{action}": {
      "post": {
        "operationId": "CompatController_partnerBookingAction",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Compat"
        ]
      }
    },
...
  },
  "info": {
    "title": "NoWayHome API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "customer@example.com"
          },
          "password": {
            "type": "string",
            "example": "StrongPass123",
            "minLength": 8
          },
          "fullName": {
            "type": "string",
            "example": "Nguyen Van A"
          },
          "phone": {
            "type": "string",
            "example": "+84901234567"
          },
          "userType": {
            "type": "string",
            "enum": [
              "customer",
              "partner",
              "admin"
            ],
            "example": "customer"
          },
          "businessName": {
            "type": "string",
            "example": "Nguyen Van A Travel Co."
          },
          "businessType": {
            "type": "string",
            "enum": [
              "individual",
              "company"
            ],
            "example": "individual"
          }
        },
        "required": [
          "email",
          "password",
          "fullName",
          "phone",
          "userType"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "customer@example.com"
          },
          "password": {
            "type": "string",
            "example": "StrongPass123",
            "minLength": 8
          },
          "deviceName": {
            "type": "string",
            "example": "Chrome on Windows"
          },
          "deviceType": {
            "type": "string",
            "enum": [
              "web",
              "ios",
              "android",
              "other"
            ],
            "example": "web"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "GoogleLoginDto": {
        "type": "object",
        "properties": {
          "credential": {
            "type": "string",
            "description": "Google credential (id_token) returned by Google Identity Services",
            "example": "eyJhbGciOiJSUzI1NiIsImtpZCI6..."
          }
        },
        "required": [
          "credential"
        ]
      },
      "UpdateKycStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "approved",
              "rejected"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
      "UpdatePropertyStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "active",
              "suspended",
              "pending_review",
              "rejected"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
      "BookingCreateReviewDto": {
        "type": "object",
        "properties": {
          "overall_rating": {
            "type": "number",
            "example": 4.5,
            "minimum": 1,
            "maximum": 5
          },
          "content": {
            "type": "string",
            "example": "Clean room and helpful service."
          }
        },
        "required": [
          "overall_rating"
        ]
      },
...
    }
  }
}